feat: Created index.html with header, page content and footer
Created a header with a logo and navigation links. Created a content section split into multiple parts meant to contain the page info; the "hero" section is what you first see upon opening the page. Created a footer with placeholder text.Created a "margin" class as a div container to easily create a "page margin" (where all text, images, etc. are aligned neatly along the page. This cannot be done by applying padding or margins to the body because it would also apply to the colored divs creating the page sections, such as the header, footer, etc.